+++ src/api/queries.ts
@@ -66,6 +66,8 @@ export const DefaultCategories = [
     '_symmetry',
     '_entity_poly',
     '_struct_asym',
+    '_struct_conn',
+    '_struct_conn_type',
     '_pdbx_struct_mod_residue',
     '_chem_comp_bond',
     '_atom_sites'
@@ -82,6 +84,8 @@ const SymmetryCategories = [
     '_struct_sheet_range',
     '_entity_poly',
     '_struct_asym',
+    '_struct_conn',
+    '_struct_conn_type',
     '_pdbx_struct_mod_residue',
     '_chem_comp_bond',
     '_atom_sites'