+++ src/lib/LiteMol-core.d.ts
@@ -1555,7 +1555,7 @@ declare namespace LiteMol.Core {
              * Checks if the computation was aborted. If so, throws.
              * Otherwise, updates the progress.
              */
-            updateProgress(msg: string, abort?: boolean | (() => void), current?: number, max?: number): void;
+            updateProgress(msg: string, abort?: boolean | (() => void), current?: number, max?: number): Promise<void>;
         }
         interface Running<A> {
             progress: Rx.Observable<Progress>;
@@ -1991,21 +1991,30 @@ declare namespace LiteMol.Core.Geometry.LinearAlgebra {
         y: number;
         z: number;
     };
+    type Matrix4 = number[];
+    type Vector3 = number[];
+    type Vector4 = number[];
     /**
      * Stores a 4x4 matrix in a column major (j * 4 + i indexing) format.
      */
     namespace Matrix4 {
-        function empty(): number[];
+        function zero(): number[];
         function identity(): number[];
+        function fromIdentity(mat: number[]): number[];
         function ofRows(rows: number[][]): number[];
         function areEqual(a: number[], b: number[], eps: number): boolean;
         function setValue(a: number[], i: number, j: number, value: number): void;
-        function copy(out: number[], a: any): number[];
+        function copy(out: number[], a: number[]): number[];
         function clone(a: number[]): number[];
         function invert(out: number[], a: number[]): number[] | null;
         function mul(out: number[], a: number[], b: number[]): number[];
+        function mul3(out: number[], a: number[], b: number[], c: number[]): number[];
         function translate(out: number[], a: number[], v: number[]): number[];
         function fromTranslation(out: number[], v: number[]): number[];
+        function rotate(out: number[], a: number[], rad: number, axis: number[]): number[] | null;
+        function fromRotation(out: number[], rad: number, axis: number[]): number[] | null;
+        function scale(out: number[], a: number[], v: number[]): number[];
+        function fromScaling(out: number[], v: number[]): number[];
         function transformVector3(out: {
             x: number;
             y: number;
@@ -2022,8 +2031,31 @@ declare namespace LiteMol.Core.Geometry.LinearAlgebra {
         function makeTable(m: number[]): string;
         function determinant(a: number[]): number;
     }
+    namespace Vector3 {
+        function obj(): ObjectVec3;
+        function zero(): number[];
+        function clone(a: number[]): number[];
+        function fromObj(v: ObjectVec3): number[];
+        function fromValues(x: number, y: number, z: number): number[];
+        function set(out: number[], x: number, y: number, z: number): number[];
+        function copy(out: number[], a: number[]): number[];
+        function add(out: number[], a: number[], b: number[]): number[];
+        function sub(out: number[], a: number[], b: number[]): number[];
+        function scale(out: number[], a: number[], b: number): number[];
+        function scaleAndAdd(out: number[], a: number[], b: number[], scale: number): number[];
+        function distance(a: number[], b: number[]): number;
+        function squaredDistance(a: number[], b: number[]): number;
+        function length(a: number[]): number;
+        function squaredLength(a: number[]): number;
+        function normalize(out: number[], a: number[]): number[];
+        function dot(a: number[], b: number[]): number;
+        function cross(out: number[], a: number[], b: number[]): number[];
+        function lerp(out: number[], a: number[], b: number[], t: number): number[];
+        function transformMat4(out: number[], a: number[], m: number[]): number[];
+        function angle(a: number[], b: number[]): number;
+    }
     namespace Vector4 {
-        function create(): number[];
+        function zero(): number[];
         function clone(a: number[]): number[];
         function fromValues(x: number, y: number, z: number, w: number): number[];
         function set(out: number[], x: number, y: number, z: number, w: number): number[];
@@ -2740,6 +2772,7 @@ declare namespace LiteMol.Core.Structure.Query {
         authSeqNumber?: number;
         insCode?: string | null;
     }
+    function allAtoms(): Builder;
     function atomsByElement(...elements: string[]): Builder;
     function atomsByName(...names: string[]): Builder;
     function atomsById(...ids: number[]): Builder;
@@ -2787,6 +2820,7 @@ declare namespace LiteMol.Core.Structure.Query {
      */
     namespace Compiler {
         function compileEverything(): (ctx: Context) => FragmentSeq;
+        function compileAllAtoms(): (ctx: Context) => FragmentSeq;
         function compileAtoms(elements: string[] | number[], sel: (model: Structure.Molecule.Model) => string[] | number[]): (ctx: Context) => FragmentSeq;
         function compileAtomIndices(indices: number[]): (ctx: Context) => FragmentSeq;
         function compileFromIndices(complement: boolean, indices: number[], tableProvider: (molecule: Structure.Molecule.Model) => {