+++ CHANGELOG.md
@@ -1,4 +1,10 @@
-﻿# 1.4.2
+﻿# 1.4.4
+* Added support for _struct_conn and _struct_conn_type categories.
+
+# 1.4.3
+* Changes were made :)
+
+# 1.4.2
 * Fixed a bug in _atom_site writer for symmetry/assembly data.
 * Optimized /ligandInteraction and /ambientResidues queries if they are used for large input sets (i.e. just /ligandInteraction).